+++ src/starter/starterstroke.c
@@ -64,7 +64,7 @@ enum auth_method_t {
 
 static char* push_string(stroke_msg_t *msg, char *string)
 {
-	u_int string_start = msg->length;
+	unsigned long string_start = msg->length;
 
 	if (string == NULL || msg->length + strlen(string) >= sizeof(stroke_msg_t))
 	{
